fix(gix-command): safely split commands and ignore shebang options
<!-- agent -->
Command preparation previously rejected every first word containing an equals
sign, so valid environment prefixes required a shell while executable names
such as tool=name could be misinterpreted as assignment-only commands. Its UTF-8
conversion also made the direct-versus-shell decision depend on encoding.

Git for Windows uses only the interpreter path from a shebang. Parsing the
suffix as shell words invented incompatible semantics and could let executable
contents supply additional interpreter options.

This provides two major fixes:

- Replace shell-words with a public byte-oriented command-line parser that
  recognizes only valid leading assignment words, applies them to the child
  environment, and preserves non-UTF-8 programs, arguments, and values during safe
  direct execution.
- Ignore all shebang options like Git for Windows, preventing executable data
  from altering the interpreter invocation.

Reuse the parser in `gix-diff` and cover its full byte-input domain with focused
unit tests and a dedicated fuzz target.

Store the required command in Outcome::command, reserve Outcome::args for
actual arguments, and reject command lines that produce no executable. Command
preparation can now consume the parser result directly without another emptiness
check or first-element extraction.

The command-line parser is consumed primarily by process-launch code, yet it
returned byte strings that every caller had to convert before constructing
`std::process::Command`.

Return OsString for the required command and its arguments so callers can use
them directly while preserving arbitrary Unix bytes. Reject values that the host
cannot represent instead of risking a conversion panic; environment assignments
remain byte-oriented for parsing.

Manual argument splitting moved leading assignments into the process environment
and spawned the parsed program directly. On Windows, Rust's lookup can fall
outside the assigned PATH and cannot launch extensionless scripts through their
shebang.

Keep assignment-prefixed commands on the shell path on Windows while retaining
manual splitting for other commands and platforms.
