+++ gix-diff/Cargo.toml
@@ -88,7 +88,6 @@ gix-filter = { path = "../gix-filter" }
 gix-traverse = { path = "../gix-traverse" }
 
 insta = "1.46.3"
-shell-words = "1"
 pretty_assertions = "1.4.0"
 
 [package.metadata.docs.rs]
